requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
aioodbc==0.4.0
pyodbc==5.0.1
aiosqlite==0.19.0
requests==2.31.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

